Check if topic exists before making a tag of it
When users attempt to input a topic an AJAX request is sent to the
Topics controller's new action - #exist, which checks whether the
wanted topic is a record in the db. If it is, a tag is made of the
topic. Otherwise, an error signals to the user that said topic does not
exist.