Prevent user from inputting a topic more than once
Topic name duplicates are no longer permitted within the hidden input
field. A custom error message is displayed if a user attempts to input
the name of a topic he already has.

Also, the html comment, which contained the previous code that was used to
input topics has been removed.