Remove most actions from the topics controller
The topics' controller functionality is fulfilled by the generated admin
topics model manager. Therefore, it is mostly obsolete. The controller's
views and resource routes have been removed.

However, the controller has been kept because of its #exists action
which determines whether a topic is present in the db. Tests have been
written to assert the proper functioning of Topic#exists.